resourceクラスの分割
