+++ lib/kakine.rb
@@ -7,6 +7,8 @@ require 'kakine/adapter'
 require 'kakine/adapter/real'
 require 'kakine/adapter/mock'
 require 'kakine/resource'
+require 'kakine/resource/openstack'
+require 'kakine/resource/yaml'
 require 'kakine/security_group'
 require 'kakine/security_rule'
 require 'kakine/validate'
